# docs/design/AI好友-OpenAPI-v1.0.yaml
openapi: 3.1.0
info:
  title: AI好友 API
  version: 1.0.0
  description: |
    AI好友 Android 客户端与 Java 后端的 MVP 契约。

    约束：
    - Base URL 为 `/api/v1`，仅允许 HTTPS；除明确标注的运维中文 HTML 页面外使用 UTF-8 JSON。
    - 时间统一使用 ISO-8601 UTC。
    - 除微信登录、邀请 proof 兑换和受限邀请会话接口外，接口默认需要 Bearer Token。
    - 写操作按接口要求携带 `Idempotency-Key`；同一主体、接口和请求摘要返回原结果，
      相同键对应不同请求摘要返回 `SESSION_CONFLICT` 或通用冲突。
    - `openId`、token、邀请 proof、OAuth code/state、邀请 Cookie、原始音频和消息正文不得进入应用日志。
    - 联系人称呼模板只用于识别发音内容，不得作为说话人身份认证或声纹识别。
    - 必须为 true 的布尔字段使用 `not: { const: false }` 表达；该约束与 `const: true` 等价，并兼容当前 Kotlin 代码生成器。
    - v1.0 使用武冈话主方言包与普通话辅助融合；任务和模板必须携带方言包及处理版本，主辅冲突时追问而非猜测。
    - 微信执行结果只表达客户端能够可信确认的状态，不代表对方已收到、已读或已接听。
  contact:
    name: AI好友项目
  x-author: codex
  x-document-date: 2026-08-27
servers:
  - url: https://api.example.com/api/v1
    description: 生产环境占位地址，发布前替换
  - url: https://api-test.example.com/api/v1
    description: 测试环境占位地址，发布前替换
...
security:
  - bearerAuth: []
paths:
...
  /contacts:
    get:
      tags: [Contacts]
      operationId: listContacts
      summary: 分页查询已绑定亲友
      description: 仅返回当前 owner 范围内的最小展示信息，不返回微信好友列表或聊天内容。
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ContactStatus"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactPageResponse"
        "401":
          $ref: "#/components/responses/AuthRequired"
...
  /contacts/{id}/local-verifications:
    post:
      tags: [Contacts]
      operationId: verifyLocalWechatContact
      summary: 完成本机微信联系人验证
      description: |
        仅通过 HTTPS 接收当前目标联系人的最小稳定定位、当前备注及验证证据；
        服务端在请求内存中完成 AES-GCM 加密和域隔离 HMAC，客户端不得自报定位摘要；
        初次验证成功后进入 ACTIVE_NO_ALIAS，录入至少一个有效称呼后才进入 ACTIVE；
        禁止上传好友列表、聊天正文、头像二进制或全屏截图。
      parameters:
        - $ref: "#/components/parameters/ContactId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocalVerificationRequest"
      responses:
        "200":
          description: 验证成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/AuthRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/WechatPageUnverified"
  /contacts/{id}/aliases:
    post:
      tags: [Contacts]
      operationId: createContactAlias
      summary: 新增方言称呼
      description: |
        两段录音必须属于当前用户和 ALIAS_ENROLLMENT 用途。唯一性在当前老人全部有效称呼模板中
        按发音内容判断，不使用 ASR 文字唯一性，也不进行说话人身份识别。两段音频、称呼记录、
        联系人状态与称呼命名空间必须原子提交；未加载经校准的方言模型和阈值时失败关闭。
      parameters:
        - $ref: "#/components/parameters/ContactId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAliasRequest"
      responses:
        "201":
          description: 称呼创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AliasResponse"
        "400":
          $ref: "#/components/responses/AudioInvalid"
        "401":
          $ref: "#/components/responses/AuthRequired"
        "403":
          $ref: "#/components/responses/ConsentRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 称呼冲突、数量达到上限或绑定状态不允许
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                phoneticConflict:
                  value:
                    code: ALIAS_PHONETIC_CONFLICT
                    message: 这个称呼与已有称呼发音太像，请换一个
                    data:
                      conflictingContactId: ct_01JEXAMPLE
                    traceId: 01JEXAMPLETRACE
                aliasLimit:
                  value:
                    code: ALIAS_LIMIT_REACHED
                    message: 该亲友最多设置 5 个称呼
                    data: null
                    traceId: 01JEXAMPLETRACE
                templateIncompatible:
                  value:
                    code: TEMPLATE_INCOMPATIBLE
                    message: 当前方言称呼模板暂不可用，请稍后重新录入
                    data: null
                    traceId: 01JEXAMPLETRACE
        "422":
          description: 两遍不一致或声学相似度处于临界区
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                inconsistent:
                  value:
                    code: ENROLLMENT_INCONSISTENT
                    message: 两遍发音不一致，请重新录两遍
                    data: null
                    traceId: 01JEXAMPLETRACE
                borderline:
                  value:
                    code: ALIAS_PHONETIC_BORDERLINE
                    message: 这个称呼发音不够容易区分，请重新录两遍
                    data: null
                    traceId: 01JEXAMPLETRACE
  /contacts/{id}/aliases/{aliasId}:
    delete:
      tags: [Contacts]
      operationId: deleteContactAlias
      summary: 删除方言称呼
      description: 删除后立即停止匹配；允许删除最后一个称呼，绑定转为 ACTIVE_NO_ALIAS。
      parameters:
        - $ref: "#/components/parameters/ContactId"
        - $ref: "#/components/parameters/AliasId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAliasRequest"
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/AuthRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /contacts/{id}:
    delete:
      tags: [Contacts]
      operationId: unbindContact
      summary: 解除亲友绑定
      description: 解除后立即从匹配和允许动作中排除，并异步清理定位、称呼模板及缓存。
      parameters:
        - $ref: "#/components/parameters/ContactId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUnbindRequest"
      responses:
        "200":
          description: 已解除
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "401":
          $ref: "#/components/responses/AuthRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 短期访问令牌；不得放入 URL、日志或崩溃报告。
    invitationSessionAuth:
      type: apiKey
      in: cookie
      name: __Host-ai_friend_invitation
      description: 固定 30 分钟的 HttpOnly、Secure、SameSite=Lax 受限邀请会话；不能登录 App 或调用其他业务接口。
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 写操作幂等键，建议使用 UUIDv7/ULID；不得复用于不同请求正文。
      schema:
        type: string
        minLength: 16
        maxLength: 128
      example: 01JEXAMPLE0000000000000000
...
    Page:
      name: page
      in: query
      required: false
      description: 页码，从 0 开始
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      description: 每页数量
      schema:
        type: integer
        minimum: 1
        maximum: 20
        default: 20
...
    ContactId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: "^ct_[A-Za-z0-9]+$"
    AliasId:
      name: aliasId
      in: path
      required: true
      schema:
        type: string
        pattern: "^al_[A-Za-z0-9]+$"
...
  responses:
    BadRequest:
      description: 请求格式或参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: VALIDATION_FAILED
            message: 请求参数不正确
            data:
              fieldErrors:
                - field: audioObjectId
                  reason: 不能为空
            traceId: 01JEXAMPLETRACE
    AuthRequired:
      description: 未登录或 token 失效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: AUTH_REQUIRED
            message: 请重新登录
            data: null
            traceId: 01JEXAMPLETRACE
...
    Forbidden:
      description: 当前主体无权操作目标资源
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ConsentRequired:
      description: 缺少本能力所需授权
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: CONSENT_REQUIRED
            message: 请先完成相关授权
            data: null
            traceId: 01JEXAMPLETRACE
    NotFound:
      description: 资源不存在或不属于当前主体
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: 状态、版本、幂等摘要或唯一约束冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    AudioInvalid:
      description: 音频格式、大小、时长、静音、噪声或有效语音不合格
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: AUDIO_INVALID
            message: 录音不清楚，请重新录制
            data: null
            traceId: 01JEXAMPLETRACE
...
    WechatPageUnverified:
      description: 客户端无法唯一确认当前微信页面/联系人，或微信与规则版本未进入白名单
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: WECHAT_PAGE_UNVERIFIED
            message: 无法确认当前联系人，请停止操作
            data: null
            traceId: 01JEXAMPLETRACE
...
  schemas:
...
    ContactStatus:
      type: string
      enum:
        - PENDING_CONSENT
        - PENDING_LOCAL_VERIFY
        - ACTIVE_NO_ALIAS
        - ACTIVE
        - REVERIFY_REQUIRED
        - REVOKED
        - BLOCKED
    AliasCompatibility:
      type: string
      enum: [COMPATIBLE, INCOMPATIBLE]
...
    ErrorResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_FAILED
            - AUTH_REQUIRED
            - DEVICE_NOT_ALLOWED
            - CONSENT_REQUIRED
            - AUDIO_INVALID
            - ENROLLMENT_INCONSISTENT
            - ALIAS_PHONETIC_CONFLICT
            - ALIAS_PHONETIC_BORDERLINE
            - CONTACT_LIMIT_REACHED
            - ALIAS_LIMIT_REACHED
            - LOCAL_VERIFICATION_REQUIRED
            - ASR_UNAVAILABLE
            - NO_CONTACT_MATCH
            - AUDIO_SEGMENT_UNCERTAIN
            - SAFETY_COMMAND_REQUIRED
            - TEMPLATE_INCOMPATIBLE
            - SESSION_CONFLICT
            - SESSION_EXPIRED
            - WECHAT_PAGE_UNVERIFIED
            - WECHAT_PARTIAL_SUCCESS
            - WECHAT_RULE_UNSUPPORTED
            - ACTION_UNSUPPORTED
            - ACCOUNT_CLOSURE_ACCEPTED
            - RATE_LIMITED
            - NOT_FOUND
            - FORBIDDEN
            - INTERNAL_ERROR
        message:
          type: string
          maxLength: 300
        data:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
        traceId:
          type: string
          minLength: 1
          maxLength: 64
      additionalProperties: false
...
    Contact:
      type: object
      required: [id, status, aliasCount, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          pattern: "^ct_[A-Za-z0-9]+$"
        displayName:
          type:
            - string
            - "null"
          maxLength: 80
          description: 当前最小展示名称，可随合法微信资料更新。
        remark:
          type:
            - string
            - "null"
          maxLength: 80
        avatarUrl:
          type:
            - string
            - "null"
          format: uri
          description: 只作展示，可刷新，不参与身份和称呼匹配。
        relationship:
          type:
            - string
            - "null"
          maxLength: 30
        status:
          $ref: "#/components/schemas/ContactStatus"
        aliasCount:
          type: integer
          minimum: 0
          maximum: 5
        aliases:
          type: array
          maxItems: 5
          items:
            $ref: "#/components/schemas/ContactAlias"
        localVerificationVersion:
          type:
            - string
            - "null"
          maxLength: 40
        verifiedAt:
          type:
            - string
            - "null"
          format: date-time
        version:
          type: integer
          format: int64
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    ContactAlias:
      type: object
      required: [id, displayText, dialectCode, dialectPackageVersion, modelVersion, thresholdVersion, compatibility, createdAt]
      properties:
        id:
          type: string
          pattern: "^al_[A-Za-z0-9]+$"
        displayText:
          type: string
          minLength: 1
          maxLength: 40
          description: 仅辅助展示，不参与唯一性判断。
        dialectCode:
          type: string
          minLength: 2
          maxLength: 40
          example: zh-Hans-CN-x-wugang
        dialectPackageVersion:
          type: string
          minLength: 1
          maxLength: 60
        modelVersion:
          type: string
          maxLength: 60
        thresholdVersion:
          type: string
          maxLength: 60
        compatibility:
          $ref: "#/components/schemas/AliasCompatibility"
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
      additionalProperties: false
    ContactPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/Contact"
        page:
          $ref: "#/components/schemas/PageMetadata"
      additionalProperties: false
    LocalVerificationRequest:
      type: object
      required:
        - stableLocator
        - pageType
        - friendConfirmed
        - locatorObservationCount
        - locatorUnique
        - wechatVersion
        - ruleVersion
        - verifiedAt
        - expectedContactVersion
      properties:
        stableLocator:
          type: string
          minLength: 1
          maxLength: 512
          writeOnly: true
          description: 当前目标联系人可唯一、可重复定位的最小稳定标识或稳定组合；仅在请求内存中短暂存在。
        currentRemark:
          type:
            - string
            - "null"
          maxLength: 80
          writeOnly: true
          description: 当前微信备注，可空；不参与联系人唯一定位。
        pageType:
          $ref: "#/components/schemas/WechatPageType"
        friendConfirmed:
          type: boolean
          description: 受支持规则是否确认目标为老人微信既有好友；false 时服务端拒绝激活。
        locatorObservationCount:
          type: integer
          minimum: 1
          maximum: 10
          description: 当前目标资料页的明确定位观察次数；MVP 只需 1 次，仍要求定位唯一且为既有好友。
        locatorUnique:
          type: boolean
          description: 当前页面内稳定定位是否唯一；false 或多候选时服务端拒绝激活。
        wechatVersion:
          type: string
          minLength: 1
          maxLength: 40
        ruleVersion:
          type: string
          minLength: 1
          maxLength: 40
        verifiedAt:
          type: string
          format: date-time
          description: 客户端完成当前页面验证的 UTC 时间；服务端只接受短时新鲜证据并以服务端时间落库。
        expectedContactVersion:
          type: integer
          format: int64
          minimum: 1
      additionalProperties: false
    WechatPageType:
      type: string
      enum: [CONTACT_PROFILE, DIRECT_CHAT, UNSUPPORTED]
    CreateAliasRequest:
      type: object
      required: [displayText, firstAudioObjectId, secondAudioObjectId, expectedContactVersion, confirmed]
      properties:
        displayText:
          type: string
          minLength: 1
          maxLength: 40
          description: 仅用于界面显示，不作为唯一键。
        phoneticHint:
          type:
            - string
            - "null"
          maxLength: 100
        firstAudioObjectId:
          type: string
          pattern: "^au_[A-Za-z0-9]+$"
        secondAudioObjectId:
          type: string
          pattern: "^au_[A-Za-z0-9]+$"
        expectedContactVersion:
          type: integer
          format: int64
          minimum: 1
        confirmed:
          type: boolean
          not:
            const: false
      additionalProperties: false
    DeleteAliasRequest:
      type: object
      required: [confirmed, expectedContactVersion]
      properties:
        confirmed:
          type: boolean
          not:
            const: false
          description: 用户已完成称呼删除二次确认。
        expectedContactVersion:
          type: integer
          format: int64
          minimum: 1
          description: 客户端当前展示的联系人对外版本；删除后该聚合版本递增。
      additionalProperties: false
...
    ContactUnbindRequest:
      type: object
      required: [confirmed, expectedContactVersion]
      properties:
        confirmed:
          type: boolean
          not:
            const: false
          description: 用户已完成解除绑定二次确认。
        expectedContactVersion:
          type: integer
          format: int64
          minimum: 1
          description: 客户端当前展示的联系人对外版本。
      additionalProperties: false
...
    ContactResponse:
      type: object
      required: [code, message, data, traceId]
      properties:
        code: {type: string, const: OK}
        message: {type: string}
        data: {$ref: "#/components/schemas/Contact"}
        traceId: {type: string}
      additionalProperties: false
    ContactPageResponse:
      type: object
      required: [code, message, data, traceId]
      properties:
        code: {type: string, const: OK}
        message: {type: string}
        data: {$ref: "#/components/schemas/ContactPage"}
        traceId: {type: string}
      additionalProperties: false
    AliasResponse:
      type: object
      required: [code, message, data, traceId]
      properties:
        code: {type: string, const: OK}
        message: {type: string}
        data: {$ref: "#/components/schemas/ContactAlias"}
        traceId: {type: string}
      additionalProperties: false
...